requirements.txt
streamlit
torch
torchvision
facenet-pytorch==2.5.2
pillow
numpy
opencv-python
pandas
